Storage addon
Added addon that contains classes for the storage and creation of reflected types.
Using the TypeStorage class you can initialize any Reflected Type using its TypeId. The class will allocate memory on the heap and initialize that memory on creation and call the destructor when TypeStorage goes out of scope.
SharedTypeStorage and WeakTypeStorage are the std::shared_ptr and std::weak_ptr variants of TypeStorage.